% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_overlap.R
\name{admission_functions}
\alias{admission_functions}
\alias{overlap_combine}
\title{WCCMH admission functions}
\usage{
overlap_combine(overlap_dt, id_cols, team_col, start_col, end_col,
  overlap_int = 1L, replace_blanks = Sys.Date(),
  priority_col = "priority")
}
\arguments{
\item{overlap_dt}{A data.table object.}

\item{id_cols}{A unique ID column(s) which will be used to group records.}

\item{team_col}{A single column; the admission record team.}

\item{start_col}{The start date for the admission record. Blanks not allowed.}

\item{end_col}{The end date for the admission record. Blanks will be
replaced, see parameter replace_blanks}

\item{overlap_int}{An integer, default value of 1, to find consecutive records.}

\item{replace_blanks}{If end_col has missing values, they will be replaced
with this value. Defaults to Sys.Date().}

\item{priority_col}{The name of the field in overlap_dt corresponding to
priority. Each team needs to have a predefined unique priority.}
}
\value{
A data.table with fixed admission records. May contain more rows
than original dataset.
}
\description{
standardize admission data
}
\note{
This could be cleaned. This could be a part II to EquaPac::overlap_fix_dt
}
\examples{
\dontrun{
input <- data.table(
person_ID = c(rep(98723, 4), rep(8534, 2), 11223, rep(22446, 3)),
team = c(rep("A", 4), rep("B", 2), "A", "B", "A", "C"),
start_date = as.Date(c("2009-10-1", "2011-11-21", "2012-1-23", "2013-3-2",
"2009-11-14", "2010-1-1", "2012-1-2", "2011-2-2", "2012-4-3", "2010-01-09")),
end_date = as.Date(c("2010-5-23", NA, "2015-01-02", "2013-3-2", "2009-12-31",
"2010-3-1", "2015-03-22", "2016-1-2", "2014-9-30", "2016-3-22")))
team_priority <- data.table(team = c("A", "B", "C"), priority = c(1, 2, 3))
input[team_priority, priority := i.priority, on = "team"]

test <- overlap_combine(overlap_dt = copy(input[team!="C"]), id_cols = "person_ID",
 team_col = "team", start_col = "start_date", end_col = "end_date",
 overlap_int = 1L, replace_blanks = Sys.Date() + 1e3,
 priority_col = "priority")
 }
}

