% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmh_colors.R
\docType{data}
\name{cmh_colors}
\alias{cmh_colors}
\alias{cmh_colors}
\title{WCCMH colors}
\format{An object of class \code{list} of length 1.}
\usage{
cmh_colors
}
\value{
A character vector of colors in some standard format.
}
\description{
A list of colors passable to ggplot2 or similar.
}
\note{
Might add more colors for different departments.
}
\examples{

cmh_colors$um_colors

}
\keyword{datasets}
