% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{convenience_strings}
\alias{convenience_strings}
\alias{add_quote}
\title{string convenience functions}
\usage{
add_quote(x, quote_char = "'")

add_quote(x, quote_char = "'")
}
\arguments{
\item{x}{character vector}

\item{quote_char}{The character(s) to quote the input vector 'x'.}
}
\value{
A quoted character vector equal to length of input x.
}
\description{
String convenience functions not in commonly used string packages.
}
\note{
Created for R-SQL automation and convenience purposes.
}
\examples{
test1 <- c("date1", "date2", "10/1/2015")
add_quote(test1)
test2 <- Sys.Date()
add_quote(test2)
}
