% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_functions.R
\name{date_functions}
\alias{date_convert}
\alias{date_functions}
\title{WCCMH date functions}
\usage{
date_convert(x, format = NULL, origin = "1970-01-01")
}
\arguments{
\item{x}{A character vector of dates.}

\item{format}{The format of the incoming date.}

\item{origin}{The origin (day 1) for the date.}
}
\value{
A vector of dates of class 'Date'.
}
\description{
standardize date classes
}
\note{
more coming soon...
}
\examples{
\dontrun{
date_convert("10/1/2015")
date_convert(x = factor("10/1/2015"))
}
}

