% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dig_dec.R
\name{dig_dec}
\alias{dig_dec}
\title{dig_dec}
\usage{
dig_dec(x)
}
\arguments{
\item{x}{Single numeric value or string that is coercible to numeric.}
}
\value{
Named numeric vector with the components
 \item{digits}{Number of digets left of the decimal}
 \item{decimals}{Number of digets right of the decimal}
 \item{round}{The smartest place to round for readablity}
}
\description{
Conviences tool to obtain a functions digits and decimals
}
\note{
Rmpfr::format, Rmpfr::mpfr may contain better solutions.
}
\section{TODO}{

\itemize{
   \item Rename dig_dec
   \item Rmpfr::format, Rmpfr::mpfr may contain better solutions.
}
}

\examples{
dig_dec(3949.1221291)
dig_dec("3523.642")
dig_dec(00000.0002)
dig_dec(x = 9.2)
}
