% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_overlap.R
\docType{data}
\name{ex_overlap}
\alias{ex_overlap}
\title{Overlap Example Dataset}
\format{A data.table object.}
\source{
Made up data, similar to CMH admission data.
}
\usage{
ex_overlap
}
\description{
An example dataset that (hopefully) incorporates all potential
problems to be fixed by overlap functions.
}
\examples{
\dontrun{
data(ex_overlap)
rm("ex_overlap")
ex_overlap
# wccmh::ex_overlap
}
}
\keyword{datasets}

