% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_data.R
\docType{data}
\name{fb_data}
\alias{fb_cpt_mod_cost}
\alias{fb_data}
\alias{fb_unit_type}
\title{Funding Bucket Data}
\format{A data.table object.}
\source{
Funding Bucket Report from PCE.
}
\usage{
fb_unit_type

fb_cpt_mod_cost
}
\description{
Encompass database is lacking unit type data for SALs,
this fixes that issue; last updated 11/3/2015.

Encompass data.table is lacking costs for SALs, this fixes that issue.
Last updated 11/23/2015.
}
\examples{
\dontrun{
data(fb_unit_type)
data(fb_cpt_mod_cost)
rm("fb_unit_type", "fb_cpt_mod_cost")
fb_unit_type
fb_cpt_mod_cost
wccmh::fb_unit_type
wccmh::fb_cpt_mod_cost
}
}
\keyword{datasets}

