% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_unit_type.R
\docType{data}
\name{fb_unit_type}
\alias{fb_unit_type}
\title{Funding Bucket Unit Type Data}
\format{A data.table object.}
\source{
Funding Bucket Report from PCE.
}
\usage{
data(fb_unit_type)
}
\description{
Encompass database is lacking unit type data for SALs, this fixes that
issue. Subject to update when necessary. Last updated 11/3/2015.
}
\examples{
\dontrun{
data(fb_unit_type)
rm("fb_unit_type")
fb_unit_type
wccmh::fb_unit_type
}
}
\keyword{datasets}

