% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal.R
\name{fiscal}
\alias{fiscal}
\alias{my_fy}
\alias{my_qtr}
\title{fiscal functions}
\usage{
my_fy(x, format = NULL, fy_start = "Oct")

my_fy(x, format = NULL, fy_start = "Oct")

my_qtr(x, format = NULL, fy_start = "Oct")
}
\arguments{
\item{x}{'Date' vector or a character vector that can be coerced to 'Date'
class.}

\item{format}{The format of the date, passed to date_convert.}

\item{fy_start}{The month the fiscal year starts. May be abbreviated to
three letters or be the full name of the month. The three-letter
abbreviation must match the package zoo's abbreviations.}
}
\value{
A character vector.
}
\description{
fiscal year, fiscal quarter, month functions for dealing with
'Date' classes.
}
\note{
Consider making a new fiscal_year class and/or fiscal_quarter class.
Also consider making some methods (this needs further consideration).
}
\examples{
my_fy("10/1/2014")
my_qtr("9/30/2015")
}

