% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fund_functions.R
\docType{data}
\name{fund_functions}
\alias{fund_functions}
\alias{recode_fund}
\alias{recode_fund_key}
\title{WCCMH fund functions}
\format{An object of class \code{list} of length 7.}
\usage{
recode_fund_key

recode_fund(x)
}
\arguments{
\item{x}{A character vector of fund names. Will be coerced to character if
class(x) is factor.}
}
\value{
A vector of recoded fund names.
}
\description{
recode_fund recodes fund character vectors.
}
\note{
more coming soon...
}
\examples{
\dontrun{
recode_fund(c("MIChild", "Medicaid"))
}

}
\keyword{datasets}

