% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nin.R
\name{\%nin\%}
\alias{\%nin\%}
\title{"Not in" for inverse value matching}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{
    a vector (numeric, character, factor)
  }

\item{table}{
    a vector (numeric, character, factor), matching the mode of \code{x}
  }
}
\value{
A logical Vector of values from x that do not have an associated
  match in y.
}
\description{
Like \%in\%, this function compares two vectors to see if any of
  the items in the left vector are not present in the right vectors,
  essentially isolating what is missing using a logicical vector.
}
\section{TODO}{

\itemize{
   \item Check to see if direct imports need to be documented like this.
}
}

\examples{
which(c(1:5,1:5,11) \%nin\% 1:10)
}
