% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosp_functions.R
\docType{data}
\name{hosp_functions}
\alias{hosp_functions}
\alias{hosp_recode_l}
\alias{recode_hosp}
\title{WCCMH hospital functions}
\format{An object of class \code{list} of length 5.}
\usage{
hosp_recode_l

recode_hosp(x)
}
\arguments{
\item{x}{A character vector of state hospital names. Will be coerced to
character if class(x) is factor.}
}
\value{
recode_state_hosp: Returns a vector of recoded state hospital names.
}
\description{
recode_state_hosp recodes state of Michigan psychiatric hospitals where
consumers are staying long-term to a standardized format.
}
\examples{
recode_hosp(c("Hawthorn Center -State Facility", "Kalamazoo Psych"))

}
\keyword{datasets}
