% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/los.R
\docType{data}
\name{los}
\alias{los}
\alias{los}
\alias{los_clinic}
\title{Length of Stay}
\format{An object of class \code{NULL} of length 0.}
\usage{
los

los_clinic(start, expiration, discharge)
}
\arguments{
\item{start}{'Date' vector or a character vector that can be coerced to
'Date' class. The start date of the psychiatric hospital admission.
Cannot be missing.}

\item{expiration}{'Date' vector or a character vector that can be coerced
to 'Date' class. The expiration date of the psychiatric hospital admission.
Cannot be missing.}

\item{discharge}{'Date' vector or a character vector that can be coerced
to 'Date' class. The discharge date of the psychiatric hospital admission.
If discharge is missing, use expiration date plus one day.}
}
\value{
A character vector.
}
\description{
los_clinic is for determining clinical, or actual, length of
stay for psychiatric hospitalizations.
paid_los is for determining the number of paid days a consumer had during a
psychiatric hospitalization.
}
\note{
Consider making a new fiscal_year class and/or fiscal_quarter class.
Also consider making some methods (this needs further consideration).
}
\examples{
\dontrun{
1+1
}
}
\keyword{datasets}
