% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/money.R
\name{money}
\alias{money}
\alias{money_add}
\alias{money_rm}
\title{Money Functions}
\usage{
money_rm(x)

money_add(x, sig_fig = 3)
}
\arguments{
\item{x}{an input vector: character for money_rm, numeric for money_add}

\item{sig_fig}{Number of significant figures; must be coerceable to
class integer and be greater than or equal to 0.}
}
\value{
A numeric vector.
}
\description{
Functions for dealing with dollar formats imported from
outside of R. Removes '$' symbol and ',' from dollar strings and
converts to numeric. Changes numbers inside of paretheses to negative.
}
\note{
Other functions for money will go in this file.
}
\examples{
money_string <- c("$1.04", "$2.24", "$2,000", "$235,658")
(no_money_string <- money_rm(money_string))
money_add(no_money_string)
money_add(x = c(1.04, 2.56e6, 3.545e9, -1.678e14))
}

