% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameRDS.R
\name{nameRDS}
\alias{nameRDS}
\alias{savenameRDS}
\alias{readnameRDS}
\title{RDS shortcuts}
\usage{
savenameRDS(x, dir)

readnameRDS(x, dir, envir = NULL)
}
\arguments{
\item{x}{R object to serialize.}

\item{dir}{The file storage location.}

\item{envir}{The environment to which readnamesRDS assign the R object. The
default is NULL in which case it reads in the R object without assigning it.}
}
\value{
all functions return either TRUE or FALSE.
}
\description{
These functions allow the user to type less when using readRDS
and saveRDS. The object name followed by RDS is appended to the file
destination.
}
\section{Notes}{

These functions are for convenience only. Does not work with \code{matrix()}
since \code{length(matrix)} does not equal 0.
}

\examples{
\dontrun{
x1 <- c(1, 2, 3)
x2 <- data.frame(v1 = 1:10)
savenameRDS(x = x1, dir = getwd())
readnameRDS(x = x1, dir = getwd())
rm(x1)
readnameRDS(x = "x1", dir = getwd(), env = .GlobalEnv)
}
}
