% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_data.R
\docType{data}
\name{overlap_data}
\alias{overlap_data}
\alias{overlap_dt}
\title{Overlap Example Dataset}
\format{A data.table object.}
\source{
Made up data, similar to CMH admission data.
}
\usage{
overlap_dt
}
\description{
An example dataset that (hopefully) incorporates all potential
problems to be fixed by overlap functions.
}
\examples{
\dontrun{
data(overlap_dt)
rm("overlap_dt")
overlap_dt
# wccmh::overlap_dt
}
}
\keyword{datasets}

