% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_functions.R
\name{overlap_functions}
\alias{overlap_combine}
\alias{overlap_functions}
\alias{priority_overlap}
\title{WCCMH admission functions}
\usage{
overlap_combine(overlap_dt, group_cols, start_col, end_col, overlap_int = 1L,
  replace_blanks = Sys.Date())

priority_overlap(overlap_dt, group_cols, priority_col, priority_value,
  start_col, end_col, overlap_int = 1L, replace_blanks = Sys.Date() + 999)
}
\arguments{
\item{overlap_dt}{A data.table object.}

\item{group_cols}{A group of columns which will collectively make a key on
which to group by when reducing/collapsing the data.table object.}

\item{start_col}{The start date for record. Blanks not allowed.}

\item{end_col}{The end date for the record. Blanks will be replaced, see
parameter replace_blanks}

\item{overlap_int}{An integer, default value of 1, to find consecutive
records.}

\item{replace_blanks}{If end_col has missing values, they will be replaced
with this value. Defaults to Sys.Date().}

\item{priority_col}{A single column which will have a priority assignment.}

\item{priority_value}{An integer column, where lower priorities override
higher priorities. May be numeric if coerceable to integer.}
}
\value{
overlap_combine returns a reduced/collapsed data.table object based
on group_cols.
priority_overlap returns a data.table with fixed records based on priority
team assignment. May contain more rows than original dataset.
}
\description{
standardize admission data
}
\note{
overlap_combine is relatively fast/optimized, although there is a for
loop that could possibly be avoided via Reduce, not sure. adm_overlap is not
code-optimized yet.
}
\examples{
# overlap_dt <- fread("C:/Users/dalrymplej/Documents/GitHub/wccmh/data/overlap_dt.csv")
# overlap_dt[, start_date := as.Date(start_date, format = '\%m/\%d/\%Y')]
# overlap_dt[, end_date := as.Date(end_date, format = "\%m/\%d/\%Y")]
# overlap_dt[, priority := as.integer(priority)]
# save(overlap_dt,
#  file = "C:/Users/dalrymplej/Documents/GitHub/wccmh/data/overlap_dt.rda")
# load("C:/Users/dalrymplej/Documents/GitHub/wccmh/data/overlap_dt.rda")
data(overlap_dt)

# how to fix if priorities are not to be accounted for...
test1 <- overlap_combine(overlap_dt = overlap_dt, group_cols = c("case_no",
"team"), start_col = "start_date", end_col = "end_date",
 overlap_int = 1L, replace_blanks = Sys.Date() + 1e3)

 # how to fix if priorities are to be accounted for...
test2 <- priority_overlap(overlap_dt = copy(overlap_dt),
                 group_cols = "person_ID",
                 priority_col = "team",
                 priority_value = "priority",
                 start_col = "start_date",
                 end_col = "end_date",
                 overlap_int = 1L,
                 replace_blanks = Sys.Date())
}

