% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_call.R
\name{p_call}
\alias{p_call}
\alias{fn}
\title{Call Holder for Function-Creation}
\usage{
p_call(..., envir = .GlobalEnv)

fn(...)
}
\arguments{
\item{...}{Accepts any R object.}

\item{envir}{The environment you want to see the final output land inside.
Default is global environment.}
}
\value{
A data.table object. Not for use, more for demonstration purposes.
}
\description{
Making data.table and call-lists, their names, and their outputs work
hand in hand. This function is for testing only and not for use in real work (yet)
}
\details{
This may or may not be a wild goose chase.
}
\section{TODO}{

\itemize{
   \item Needs review, documentation, and rewrite.
}
}

\examples{
\dontrun{
DT <- data.table(1:10, letters[1:10])
letter_l <- as.list(LETTERS[1:10])
# i cant seem to save this print statement, but w/e not really needing to either
p_call("Hey", "Look!", DT, "dont try this at home", letter_l, "but", "have fun")
dt_call[call_char=="p_call", call_in]
dt_call[call_char=="DT", output]
vec_string <- c("this", "is", "a", "string")
vec_numeric <- c(1, 1, 2, 3, 5, 8, 13)
vec_bool <- c(TRUE, FALSE, FALSE, TRUE)


p_call(DT, vec, "phrase")
p_call("long string with \\n characters to for the return line")
p_call("a mixture of every conceivable usage case")

}


}
