% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parity_functions.R
\name{parity_functions}
\alias{even_char}
\alias{even_obj}
\alias{odd_char}
\alias{odd_obj}
\alias{parity_functions}
\title{Even/Odd (Parity) Functions}
\usage{
odd_char(x)

even_char(x)

even_obj(x)

odd_obj(x)
}
\arguments{
\item{x}{A character vector for even_char/odd_char or a R list/vector for
even_obj/odd_obj.}
}
\value{
the even or odd elements of the input, in vector format for odd_char/even_char,
and the format remains consistent with the input for even_obj/odd_obj.
}
\description{
Extract even/odd characters via odd_char/even_char. Extract
even/odd elements of an R objects (vector or list) via even_obj/odd_obj
}
\note{
convenience functions only. even_char/odd_char perhaps could be more
efficient.
}
\examples{
odd_char(paste(letters[1:26], collapse = ""))
even_char("a1b2c3d4")
ex_list <-
 list(a1 = "warthog", b2 = "bomber",
 c3 = "PO", c4 = "explosives are dangerous")
odd_obj(ex_list)
even_obj(ex_list)

}

