% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.fb.R
\docType{data}
\name{read.fb}
\alias{char_cols}
\alias{read.fb}
\alias{read_header}
\title{FB file reader}
\format{\preformatted{List of 52
 $ PROVIDER NAME       : chr "character"
 $ PROVIDER ID         : chr "integer"
 $ CASE #              : chr "integer"
 $ PRI PROCEDURE CODE  : chr "character"
 $ SEC PROCECURE CODE  : chr [1:2] "logical" "character"
 $ MOD                 : chr "character"
 $ SERVICE DESC        : chr "character"
 $ CATEGORY DESC       : chr "character"
 $ UNIT TYPE           : chr "character"
 $ FROM DATE           : chr "character"
 $ THRU DATE           : chr "character"
 $ UNITS               : chr "integer"
 $ DURATION            : chr "integer"
 $ ALLOWED AMOUNT      : chr "numeric"
 $ CHARGED AMOUNT      : chr "integer"
 $ POPULATION          : chr "character"
 $ ADULT OR CHILD      : chr "character"
 $ AGE                 : chr "integer"
 $ COV: MEDICAID       : chr "character"
 $ COV: HAB WAIVER     : chr "character"
 $ COV: CHILD WAIVER   : chr "character"
 $ COV: SED WAIVER     : chr "character"
 $ COV: ABW            : chr "character"
 $ COV: HEALTHY MI PLAN: chr "character"
 $ COV: MI CHILD       : chr "character"
 $ COV: AUTISM         : chr "character"
 $ COV: COFR           : chr "character"
 $ SRV: STATE PLAN     : chr "character"
 $ SRV: ALT/B3         : chr "character"
 $ SRV: HAB WAIVER     : chr "character"
 $ SRV: EPSDT          : chr "character"
 $ SRV: CHILD WAIVER   : chr "character"
 $ SRV: GF ONLY        : chr "character"
 $ BKT: GF             : chr "character"
 $ BKT: STATE PLAN     : chr "character"
 $ BKT: ALT/B3         : chr "character"
 $ BKT: EPSDT          : chr "character"
 $ BKT: HAB WAIVER     : chr "character"
 $ BKT: ABW            : chr "character"
 $ BKT: HEALTHY MI PLAN: chr "character"
 $ BKT: MICHILD        : chr "character"
 $ BKT: CHILD WAIVER   : chr "character"
 $ BKT: SED WAIVER     : chr "character"
 $ BKT: AUTISM/MICHILD : chr "character"
 $ BKT: AUTISM/MEDICAID: chr "character"
 $ BKT: COFR           : chr "character"
 $ STAFF NAME          : chr "character"
 $ PLACE OF SERVICE    : chr "integer"
 $ BATCH ID            : chr "integer"
 $ CLAIM ID            : chr "integer"
 $ CLAIM DETAIL ID     : chr "integer"
 $ SOURCE              : chr "character"
}}
\usage{
read.fb(file_path)
}
\arguments{
\item{file_path}{Path to fb .csv file.}
}
\value{
A data.table object that is also class 'fb.'
}
\description{
FB file reader

Read in funding bucket files via data.table::fread and convert
them to the "fb" class. The header is stored as attributes.
Applies attrubutes based off of the file header information.
}
\note{
Might even format if James isn't to lazy to do it himself.
}
\examples{
\dontrun{
require(data.table)
folder_path  <- 'C:/Users/dalrymplej/Dropbox/EToPac/data-sets'
file_name <- 'fb contract 10_01_14_to_07_31_15 run 10_19_15.csv'
file_name <- 'fb direct 10_01_14_to_07_31_15 run 10_19_15.csv'
file_path <- file.path(folder_path, file_name)
my_fb_dt <- read.fb(file_path)
lapply(my_fb_dt, class)
attributes(my_fb_dt)
}
}
\keyword{datasets}

