% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_functions.R
\docType{data}
\name{team_functions}
\alias{cmh_priority_dt}
\alias{cmh_program_key}
\alias{cmh_recode}
\alias{cmh_team_key}
\alias{cmh_teams_f}
\alias{recode_team_prog}
\alias{team_functions}
\title{WCCMH team functions}
\format{An object of class \code{list} of length 10.}
\usage{
cmh_team_key

cmh_program_key

cmh_recode(x)

recode_team_prog(x)

cmh_teams_f(x, levels = c("ACT", "DD", "MI", "Child HB", "Child"),
  level_order = NULL)

cmh_priority_dt
}
\arguments{
\item{x}{A character vector of team names. Will be coerced to character if
class(x) is factor.}

\item{levels}{The levels that will be assigned. Unspecified inputs result
in NA.}

\item{level_order}{The order of the levels. Defaults to NULL.}
}
\value{
A vector of recoded team names.
}
\description{
cmh_recode recodes Washtenaw CMH team names to a standardized
team format.
recode_team_prog recodes Washtenaw CMH team/program names to a standardized
program format.
cmh_teams_f factors (ordered is an option) teams.
cmh_priority_dt assigns a priority to all of the main teams.
}
\note{
need testing, consider adding an automatic "missing" assignment
with a warning message.
}
\examples{
\dontrun{
cmh_recode("WSH - ACT")
cmh_recode(c("WSH - ACT", "DD Adult"))
}
}
\keyword{datasets}

