% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{word}
\alias{word}
\alias{single_cap}
\alias{cap_word}
\alias{locus_word2num}
\title{WCCMH word functions}
\usage{
single_cap(x)

cap_word(x)

locus_word2num(x)
}
\arguments{
\item{x}{A character vector which will be capitalized via sapply.}
}
\value{
A vector of recoded fund names.
}
\description{
cap_word capitalizes the first letter of every
space-separated word.
}
\note{
more coming soon...
}
\examples{
cap_word(c("level", "one"))
cap_word(c("level one", "one", "this is a sentence."))
locus_word2num("Level One")
}
