% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cmh_colors.R
\docType{data}
\name{cmh_colors}
\alias{cmh_colors}
\title{WCCMH colors}
\format{\preformatted{List of 1
 $ um: chr [1:5] "#94B6D2" "#DD8047" "#EBDDC3" "#775F55" ...
}}
\usage{
cmh_colors
}
\value{
A character vector of colors in some standard format.
}
\description{
WCCMH colors

A list of colors passable to ggplot2 or similar.
}
\note{
Might add more colors for different departments.
}
\examples{
cmh_colors$um_colors
}
\keyword{datasets}

