% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/money.R
\name{money}
\alias{money}
\alias{money_add}
\alias{money_add_one}
\alias{money_rm}
\title{Money Functions}
\usage{
money_rm(x)

money_add(x)
}
\arguments{
\item{x}{an input vector: character for money_rm, numeric for money_add}
}
\value{
A numeric vector.
}
\description{
Money Functions

Functions for dealing with dollar formats imported from
outside of R. Removes '$' symbol and ',' from dollar strings and
converts to numeric. Changes numbers inside of paretheses to negative.
}
\note{
Other functions for money will go in this file.
}
\examples{
money_string <- c("$1.04", "$2.24")
(no_money_string <- money_rm(money_string))
money_add(no_money_string)
}

