% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.fb.R
\docType{data}
\name{read.fb}
\alias{read.fb}
\alias{char_cols}
\alias{read_header}
\alias{read.fb}
\title{FB file reader}
\format{An object of class \code{list} of length 52.}
\usage{
char_cols

read_header(file_path)

read.fb(file_path)
}
\arguments{
\item{file_path}{Path to fb .csv file.}
}
\value{
A data.table object that is also class 'fb.'
}
\description{
Read in funding bucket files via data.table::fread and convert
them to the "fb" class. The header is stored as attributes. Applies
attrubutes based off of the file header information.
}
\note{
Could format the columns coming in as factors if that were desirable.
}
\examples{
\dontrun{
require(data.table)
folder_path  <- 'C:/Users/dalrymplej/Dropbox/EToPac/data-sets'
file_name <- 'fb contract 10_01_14_to_07_31_15 run 10_19_15.csv'
file_name <- 'fb direct 10_01_14_to_07_31_15 run 10_19_15.csv'
file_path <- file.path(folder_path, file_name)
my_fb_dt <- read.fb(file_path)
lapply(my_fb_dt, class)
attributes(my_fb_dt)
}
}
\keyword{datasets}
